# api/openapi/auth.yml
openapi: 3.0.1
info:
  title: Mainflux authentication service
  description: HTTP API for managing platform API keys.
  version: "1.0.0"
paths:
...
  /orgs/{orgId}/memberships/{memberId}:
    get:
      summary: Retrieves organization membership details.
      description: |
        Retrieves membership details for the given organization ID and member ID.
      tags:
        - auth
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/MemberId"
      responses:
        '200':
          $ref: "#/components/responses/OrgMembershipRes"
        '400':
          description: Failed due to malformed query parameters.
        '401':
          description: Missing or invalid access token provided.
        '404':
          description: Organization or member does not exist.
        '500':
          $ref: "#/components/responses/ServiceError"
  /orgs/{orgId}/memberships:
    post:
      summary: Creates memberships for organization.
      description: |
        Creates new memberships for organization identified by the given ID.
      tags:
        - auth
      parameters:
        - $ref: "#/components/parameters/OrgId"
      requestBody:
        $ref: "#/components/requestBodies/OrgMembershipsReq"
      responses:
        '200':
          description: Memberships created for organization.
        '400':
          description: Failed due to malformed JSON.
        '401':
          description: Missing or invalid access token provided.
        '404':
          description: Failed due to non existing organization.
        '409':
          description: Failed due to using already existing ID.
        '500':
          $ref: "#/components/responses/ServiceError"
    get:
      summary: Retrieves all memberships of organization.
      description: |
        Retrieves all memberships of organization identified by the given ID.
      tags:
        - auth
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Metadata"
      responses:
        '200':
          $ref: "#/components/responses/OrgMembershipsRes"
        '400':
          description: Failed due to malformed query parameters.
        '401':
          description: Missing or invalid access token provided.
        '415':
          description: Missing or invalid content type.
        '500':
          $ref: "#/components/responses/ServiceError"
    patch:
      summary: Removes memberships from organization.
      description: |
        Removes memberships from organization identified by the given ID.
      tags:
        - auth
      parameters:
        - $ref: "#/components/parameters/OrgId"
      requestBody:
        $ref: "#/components/requestBodies/RemoveOrgMembershipsReq"
      responses:
        '204':
          description: Memberships removed from organization.
        '400':
          description: Failed due to malformed JSON.
        '401':
          description: Missing or invalid access token provided.
        '500':
          $ref: "#/components/responses/ServiceError"
    put:
      summary: Updates memberships of organization.
      description: |
        Updates members role of organization identified by the given ID.
      tags:
        - auth
      parameters:
        - $ref: "#/components/parameters/OrgId"
      requestBody:
        $ref: "#/components/requestBodies/OrgMembershipsReq"
      responses:
        '200':
          description: Membership updated.
        '400':
          description: Failed due to malformed JSON.
        '401':
          description: Missing or invalid access token provided.
        '404':
          description: Failed due to non existing organization.
        '500':
          $ref: "#/components/responses/ServiceError"
  /orgs/{orgId}/memberships/backup:
    get:
      summary: Retrieves organization membership backup.
      description: |
        Retrieves a backup of all organization memberships. The response is a JSON object containing all related organization memberships. 
      tags:
        - auth
      parameters:
        - $ref: "#/components/parameters/OrgId"
      responses:
        '200':
          $ref: "#/components/responses/OrgMembershipsBackupRes"
        '401':
          description: Missing or invalid access token provided.
        '403':
          description: Failed to perform authorization over the entity.
        '404':
          description: Failed due to non existing organization or related entities.
        '422':
          description: Database can't process request.
        '500':
          $ref: "#/components/responses/ServiceError"
  /orgs/{orgId}/memberships/restore:
    post:
      summary: Restores organization memberships backup.
      description: |
        Restores organization memberships from backup. Backup is a JSON file that contains all organization memberships that will be restored.
      tags:
        - auth
      parameters:
        - $ref: "#/components/parameters/OrgId"
      requestBody:
        $ref: "#/components/requestBodies/RestoreOrgMembershipsReq"
      responses:
        '201':
          description: Backup restored.
        '400':
          description: Failed due to malformed JSON.
        '401':
          description: Missing or invalid access token provided.
        '415':
          description: Missing or invalid content type.
        '500':
          $ref: "#/components/responses/ServiceError"
...
components:
  schemas:
...
    OrgMembership:
      type: object
      properties:
        member_id:
          type: string
          format: uuid
        email:
          type: string
          example: "test@example.com"
          description: Organization member email.
        role:
          type: string
          example: "viewer"
          description: Organization member role.
    OrgMembershipPageSchema:
      type: object
      properties:
          limit:
            type: integer
            description: Limit of the page.
          offset:
            type: integer
            description: Offset of the page.
          total:
            type: integer
            description: Total number of members.
          name:
            type: string
          org_memberships:
            type: array
            minItems: 1
            uniqueItems: true
            items:
              $ref: "#/components/schemas/OrgMembership"
    OrgMembershipsSchema:
      type: object
      properties:
        org_memberships:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: "#/components/schemas/OrgMembership"
    OrgMemberIDsSchema:
      type: object
      properties:
        member_ids:
          type: array
          example: ["c5747f2f-2a7c-4fe1-b41a-51a5ae290945"]
          minItems: 1
          uniqueItems: true
          items:
            type: string
            format: uuid
    OrgMembershipsResSchema:
      type: object
      properties:
        member_id:
          type: string
          format: uuid
          description: Unique member identifier generated by the service
        org_id:
          type: string
          format: uuid
          description: Unique organization identifier generated by the service
        role:
          type: string
          description: Member role in the organization
        created_at:
          type: string
          description: Time when the member relation is created.
        updated_at:
          type: string
          description: Time when the member relation is updated.
      required:
        - member_id
        - org_id
        - role
        - created_at
        - updated_at
...
    BackupAndRestoreOrgMembershipsSchema:
      type: array
      minItems: 0
      uniqueItems: true
      items:
        $ref: "#/components/schemas/OrgMembershipsResSchema"
...
  parameters:
...
    OrgId:
      name: orgId
      description: Organization ID.
      in: path
      schema:
        type: string
        format: uuid
      required: true
...
    MemberId:
      name: memberId
      description: Member ID.
      in: path
      schema:
        type: string
        format: uuid
      required: true
    Metadata:
      name: metadata
      description: Metadata filter. Filtering is performed matching the parameter with metadata on top level. Parameter is json.
      in: query
      schema:
        type: string
        minimum: 0
      required: false
    Limit:
      name: limit
      description: Size of the subset to retrieve.
      in: query
      schema:
        type: integer
        default: 10
        minimum: 1
      required: false
    Offset:
      name: offset
      description: Number of items to skip during retrieval.
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
      required: false
  requestBodies:
...
    OrgMembershipsReq:
      description: JSON-formatted document describing a request to create and update memberships.
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/OrgMembershipsSchema"
    RestoreOrgMembershipsReq:
      description: Binary-encoded JSON file describing restore organization memberships request.
      required: true
      content:
        application/octet-stream:
          schema:
            $ref: "#/components/schemas/BackupAndRestoreOrgMembershipsSchema"
    RemoveOrgMembershipsReq:
      description: JSON-formatted document describing a membership removal request.
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/OrgMemberIDsSchema"
...
  responses:
    ServiceError:
      description: Unexpected server-side error occurred.
...
    OrgMembershipsRes:
      description: Data retrieved.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/OrgMembershipPageSchema"
    OrgMembershipRes:
      description: Data retrieved.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/OrgMembership"
    OrgMembershipsBackupRes:
      description: Successfully retrieved organization membership backup.
      content:
        application/octet-stream:
          schema:
            $ref: "#/components/schemas/BackupAndRestoreOrgMembershipsSchema"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        * Users access: "Authorization: Bearer <user_token>"
security:
  - bearerAuth: []